==> RMD/1 - Analisis de Supervivencia y Machine Learning/Codigo fuente/Rmds/10_Configuracion_Inicial.Rmd <==
<br>
<br>

# Configuración inicial {-}

<br>

```{r, include = FALSE, warning=FALSE}
library(knitr)
knit_hooks_ORIGINAL <- knit_hooks

local({

    hook_plot <- knit_hooks$get('plot')
    knit_hooks$set(plot = function(x, options) {
      x <- paste(x, collapse = '.')
      if (!grepl('\\.svg', x)) return(hook_plot(x, options))
      # read the content of the svg image and write it out without <?xml ... ?>
      paste(readLines(x)[-1], collapse = '\n')
    })

})

```

```{r, include = FALSE, warning=FALSE}

```

##### Cargamos las librerías que son necesarias {-}

```{r Librerias, warning=FALSE}
suppressMessages(library(RTCGAToolbox)) # Datos utilizados (The Cancer Genome Atlas)
suppressMessages(library(dplyr))        # Librería usada para la manipulación de tablas
suppressMessages(library(kableExtra))   # Librería usada para generar tablas
suppressMessages(library(survival))     # Librería base  para  análisis  de supervivencia
suppressMessages(library(survminer))    # Librería usada para las curvas de supervivencia
suppressMessages(library(limma))        # Librería usada para Análisis de datos y modelos lineales
suppressMessages(library(reticulate))   # Librería usada para usar el motor de python
```

```{r, include = FALSE, warning=FALSE}
suppressMessages(library(RColorBrewer))
```


<br>

##### Cargamos los datos necesarios {-}

TCGAretriever es un paquete de R diseñado para facilitar la descarga de datos de cáncer de The Cancer Genome Atlas (TCGA). Este paquete permite a los usuarios descargar datos genéticos y clínicos de pacientes con cáncer de una variedad de tipos, incluyendo cáncer de mama.

Hay que cargar los datos desde un archivo usando `load("../brcaData.RData")` ya que no se pueden descargar utilizando TCGAretriever directamente en windows. El archivo contiene información genética y clínica de pacientes con cáncer de mama.

```{r Carga_de_datos}
load("../brcaData.RData")

brcaData.cl <- getData(brcaData, "clinical")                # Datos Clínicos

brcaData.gene <- brcaData@RNASeq2GeneNorm[[1]]@DataMatrix   # Datos Genómicos
```

<br>

##### Modificamos los tipos de datos {-}

Modificar los datos que vienen por defecto es necesario para poder visualizarlos de manera adecuada y para poder realizar análisis estadísticos en ellos. Los datos pueden venir en un formato que no permita la visualización o análisis, como por ejemplo que sea `character` en lugar de un valor numérico o un factor.

Además, algunos gráficos y paquetes de R requieren que los datos estén en un formato específico para poder mostrarlos de manera adecuada

```{r}
# Conjunto de datos numéricos
brcaData.cl$years_to_birth <- as.double(brcaData.cl$years_to_birth)
brcaData.cl$days_to_death <- as.double(brcaData.cl$days_to_death)
brcaData.cl$days_to_last_followup <- as.double(brcaData.cl$days_to_last_followup)
brcaData.cl$date_of_initial_pathologic_diagnosis <- as.integer(brcaData.cl$date_of_initial_pathologic_diagnosis)
brcaData.cl$number_of_lymph_nodes <- as.double(brcaData.cl$number_of_lymph_nodes)


# Conjunto de datos lógicos
brcaData.cl$vital_status <- as.logical(as.integer(brcaData.cl$vital_status))
brcaData.cl$radiation_therapy <- ifelse(brcaData.cl$radiation_therapy=="yes",1,0)
brcaData.cl$radiation_therapy <- as.logical(as.integer(brcaData.cl$radiation_therapy))


# Conjunto de datos de tipo factor
brcaData.cl$`Composite Element REF` <- as.factor(brcaData.cl$`Composite Element REF`)
brcaData.cl$tumor_tissue_site <- as.factor(brcaData.cl$tumor_tissue_site)
brcaData.cl$pathologic_stage <- as.factor(brcaData.cl$pathologic_stage)
brcaData.cl$pathology_T_stage <- as.factor(brcaData.cl$pathology_T_stage)
brcaData.cl$pathology_N_stage <- as.factor(brcaData.cl$pathology_N_stage)
brcaData.cl$pathology_M_stage <- as.factor(brcaData.cl$pathology_M_stage)
brcaData.cl$gender <- as.factor(brcaData.cl$gender)
brcaData.cl$histological_type <- as.factor(brcaData.cl$histological_type)
brcaData.cl$race <- as.factor(brcaData.cl$race)
brcaData.cl$ethnicity <- as.factor(brcaData.cl$ethnicity)
```

<br>

<!-- SkimR Puede servir también  -->

##  Visualizamos el resumen de los datos {.tabset .tabset-fade .tabset-pills -}

`r params$NOTE`

<center>Como hay distintos tipos de datos, las métricas de cada uno se visualizan mejor por separado. Además, para no tener que reutilizar tantas veces el mismo código, defino la siguiente función para mostrar las tablas: </center>

<br>

```{r}
simple_table <- function(df) {
  knitr::kable(summary(df)) %>% kable_minimal()
}
```


</div>

### Double {-}
```{r}
brcaData.cl.double <- brcaData.cl[c("years_to_birth", "days_to_death", "days_to_last_followup", "date_of_initial_pathologic_diagnosis", "number_of_lymph_nodes")]

simple_table(brcaData.cl.double)
```

<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>

### Logical {-}
```{r}
brcaData.cl.logical <- brcaData.cl[c("vital_status", "radiation_therapy")]

simple_table(brcaData.cl.logical)
```

<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>
<br>

### Factor {-}

<center>
`r params$NOTE` Para visualizar correctamente las tablas segun el numero de factores separamos en distintas secciones </div>
</center>

###### Tablas de un solo factor {-}
<details style="background-color: azure; border: 2px solid azure1; border-radius: 5px; padding: 10px 10px 10px 10px;">
<summary markdown="span" style="background-color: azure; color: #398dd1; padding: 3px 2px 3px 2px;"> Informacion de la tabla </summary>
```{r}
brcaData.cl.factor <- brcaData.cl[c("Composite Element REF", "tumor_tissue_site", "pathologic_stage", "pathology_T_stage", "pathology_N_stage", "pathology_M_stage", "gender", "histological_type", "race", "ethnicity")]

colnames <- c("Composite Element REF","tumor_tissue_site") ## Identificadores
simple_table(brcaData.cl.factor[colnames])
```
</details>

###### Tabla de dos factores {-}
<details style="background-color: azure; border: 2px solid azure1; border-radius: 5px; padding: 10px 10px 10px 10px;">
<summary markdown="span" style="background-color: azure; color: #398dd1; padding: 3px 2px 3px 2px;"> Informacion de la tabla </summary>
```{r}
simple_table(brcaData.cl.factor[c("gender")])
```
</details>

###### Tablas entre tres y cinco factores {-}
<details style="background-color: azure; border: 2px solid azure1; border-radius: 5px; padding: 10px 10px 10px 10px;">
<summary markdown="span" style="background-color: azure; color: #398dd1; padding: 3px 2px 3px 2px;"> Informacion de la tabla </summary>
```{r}
simple_table(brcaData.cl.factor[c("pathology_M_stage", "race", "ethnicity")])
```
</details>

###### Tablas varios factores {-}
<details style="background-color: azure; border: 2px solid azure1; border-radius: 5px; padding: 10px 10px 10px 10px;">
<summary markdown="span" style="background-color: azure; color: #398dd1; padding: 3px 2px 3px 2px;"> Informacion de la tabla </summary>
```{r}
simple_table(brcaData.cl.factor[c("histological_type", "pathologic_stage","pathology_T_stage","pathology_N_stage")])
```
</details>

<br>
<br>
<br>
